script: Move CSP DOM interfaces to `script/dom/security`
