+++ components/script/dom/security/securitypolicyviolationevent.rs
@@ -6,13 +6,13 @@ use dom_struct::dom_struct;
 use js::rust::HandleObject;
 use stylo_atoms::Atom;
 
-use super::bindings::reflector::reflect_dom_object_with_proto;
 use crate::dom::bindings::codegen::Bindings::EventBinding::Event_Binding::EventMethods;
 use crate::dom::bindings::codegen::Bindings::SecurityPolicyViolationEventBinding::{
     SecurityPolicyViolationEventDisposition, SecurityPolicyViolationEventInit,
     SecurityPolicyViolationEventMethods,
 };
 use crate::dom::bindings::inheritance::Castable;
+use crate::dom::bindings::reflector::reflect_dom_object_with_proto;
 use crate::dom::bindings::root::DomRoot;
 use crate::dom::bindings::str::{DOMString, USVString};
 use crate::dom::event::{Event, EventBubbles, EventCancelable, EventComposed};
